requirements.txt
Flask
flask_socketio
openai
python-dotenv
requests
selenium
webdriver-manager
